# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CTV Live Events API",
    "version": "v1",
    "description": "Local MVP API for live sports discovery and state refresh."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "paths": {
...
    "/api/v1/events/upcoming": {
      "get": {
        "summary": "Return upcoming matches with context and pre-match intelligence",
        "security": [
          {
            "ApiKeyBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Upcoming events"
          },
          "503": {
            "description": "AI usage disabled"
          }
        }
      }
    },
    "/api/v1/events/upcoming/{matchId}": {
      "get": {
        "summary": "Return one upcoming match by match ID",
        "security": [
          {
            "ApiKeyBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Upcoming event"
          },
          "404": {
            "description": "Upcoming match not found"
          }
        }
      }
    },
...
  }
}